+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!--next-version-placeholder-->
 
+## [v3.6.7-dev3] - 2024-09-09
+### Changed
+* Testing pydicom 3.0.0.
+* Drop support for python 3.9.
+
+## [v3.6.7-dev2] - 2024-09-05
+### Added
+* DICOMPlugin: Read non-image datasets (e.g. structured reports) also.
+
+## [v3.6.7-dev1] - 2024-09-02
+### Added
+* Read DICOM image and non-image datasets separately.
+
 ## [v3.6.7-dev0] - 2024-08-08
 ### Fixed
 * DICOMPlugin.__get_transformation_matrix(): Fixed problem when there is one slice only.